Create cache:context:list command and move cache:rebuild command into new namespace
