+++ config/services/cache.yml
@@ -1,0 +1,9 @@
+services:
+  cache_context_list:
+    class: Drupal\Console\Command\Cache\CacheContextListCommand
+    tags:
+      - { name: console.command }
+  cache_rebuild:
+    class: Drupal\Console\Command\Cache\CacheRebuildCommand
+    tags:
+      - { name: console.command }
