+++ config/translations/en/cache.context.list.yml
@@ -1,0 +1,5 @@
+description: 'Displays current cache context for the application.'
+messages:
+    code: 'Context ID'
+    label: 'Label'
+    class: 'Class path'
